Adds Facet suffix to Algeron facets
