+++ src/main/java/org/jboss/forge/arquillian/command/algeron/publisher/AbstractAlgeronPublisherCommand.java
@@ -11,7 +11,7 @@ import org.jboss.forge.addon.ui.context.UIExecutionContext;
 import org.jboss.forge.addon.ui.result.Result;
 import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.arquillian.api.algeron.AlgeronPublisherFacet;
-import org.jboss.forge.arquillian.api.algeron.AlgeronConsumer;
+import org.jboss.forge.arquillian.api.algeron.AlgeronConsumerFacet;
 
 import javax.inject.Inject;
 import java.util.Map;
@@ -63,7 +63,7 @@ public abstract class AbstractAlgeronPublisherCommand extends AbstractProjectCom
     public boolean isEnabled(UIContext context) {
         Boolean parent = super.isEnabled(context);
         if (parent) {
-            return getSelectedProject(context).hasFacet(AlgeronConsumer.class);
+            return getSelectedProject(context).hasFacet(AlgeronConsumerFacet.class);
         }
         return parent;
     }
